requirements.pinned.txt
# -*- coding: utf-8 -*-
#
# Copyright (C) 2020 CERN.
#
# CDS-ILS is free software; you can redistribute it and/or modify it under
# the terms of the MIT License; see LICENSE file for more details.

alembic==1.4.3
amqp==5.0.2
arrow==0.17.0
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
base32-lib==1.0.2
billiard==3.6.3.0
bleach==3.2.1
blinker==1.4
cachelib==0.1.1
cds-dojson==0.9.0
celery==5.0.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-default-group==1.2.2
click-didyoumean==0.0.3
click-repl==0.1.6
counter-robots==2018.6
cryptography==3.2.1
decorator==4.4.2
dnspython==2.0.0
dojson==1.4.0
elasticsearch==7.10.0
elasticsearch-dsl==7.3.0
email-validator==1.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Alembic==2.0.1
Flask-Assets==2.0
Flask-BabelEx==0.9.4
Flask-Breadcrumbs==0.5.1
Flask-Caching==1.9.0
Flask-CeleryExt==0.3.4
Flask-Collect==1.2.2
Flask-Cors==3.0.9
Flask-DebugToolbar==0.11.0
Flask-KVSession-Invenio==0.6.3
Flask-Limiter==1.1.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Menu==0.7.2
Flask-OAuthlib==0.9.6
Flask-Principal==0.4.0
Flask-Security==3.0.0
flask-shell-ipython==0.4.1
Flask-SQLAlchemy==2.4.4
flask-talisman==0.5.0
flask-webpackext==1.0.2
Flask-WTF==0.14.3
fs==0.5.4
ftfy==4.4.3
future==0.18.2
fuzzywuzzy==0.18.0
html5lib==1.1
idna==2.10
importlib-metadata==2.0.0
infinity==1.5
iniconfig==1.0.1
intervals==0.9.0
invenio==3.3.0
invenio-access==1.4.1
invenio-accounts==1.3.1
invenio-admin==1.2.1
invenio-app==1.2.7
invenio-app-ils==1.0.0a20
invenio-assets==1.1.5
invenio-banners==1.0.0a1
invenio-base==1.2.3
invenio-cache==1.1.0
invenio-celery==1.2.1
invenio-circulation==1.0.0a28
invenio-config==1.0.3
invenio-db==1.0.7
invenio-files-rest==1.2.0
invenio-formatter==1.0.3
invenio-i18n==1.2.0
invenio-indexer==1.1.2
invenio-jsonschemas==1.1.0
invenio-logging==1.3.0
invenio-mail==1.0.2
invenio-migrator==1.0.0a10
invenio-oaiserver==1.2.0
invenio-oauth2server==1.2.0
invenio-oauthclient==1.3.5
invenio-opendefinition==1.0.0a9
invenio-pages==1.0.0a5
invenio-pidrelations==1.0.0a7
invenio-pidstore==1.2.1
invenio-query-parser==0.6.0
invenio-queues==1.0.0a3
invenio-records==1.3.2
invenio-records-rest==1.7.2
invenio-rest==1.2.2
invenio-search==1.3.1
invenio-stats==1.0.0a18
invenio-theme==1.1.4
invenio-userprofiles==1.1.1
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsmin==2.2.2
jsonpatch==1.26
jsonpointer==2.0
jsonref==0.2
jsonresolver==0.3.1
jsonschema==3.2.0
kombu==5.0.2
limits==1.5.1
lorem==0.1.1
lxml==4.6.1
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.9.1
maxminddb==2.0.3
maxminddb-geolite2==2018.703
msgpack==1.0.0
node-semver==0.1.1
oauthlib==2.1.0
ordereddict==1.1
packaging==20.4
parso==0.7.1
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pluggy==0.13.1
prompt-toolkit==3.0.8
psycopg2-binary==2.8.6
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.20
Pygments==2.7.2
PyJWT==1.7.1
pynpm==0.1.2
pyparsing==2.4.7
pyPEG2==2.15.2
pyrsistent==0.17.3
pytest-celery==0.0.0a1
python-dateutil==2.8.1
python-editor==1.0.4
python-geoip==1.2
python-ldap==3.2.0
pytz==2020.4
pywebpack==1.1.0
PyYAML==5.3.1
redis==3.5.3
requests==2.25.0
requests-oauthlib==1.1.0
sentry-sdk==0.19.3
simplejson==3.17.2
simplekv==0.14.1
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.20
SQLAlchemy-Continuum==1.3.11
SQLAlchemy-Utils==0.35.0
traitlets==4.3.3
ua-parser==0.10.0
uritools==3.0.0
urllib3==1.26.2
uWSGI==2.0.19.1
uwsgi-tools==1.1.1
uwsgitop==0.11
validators==0.18.1
vine==5.0.0
wcwidth==0.2.5
webargs==5.5.3
webassets==2.0
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
WTForms-Alchemy==0.17.0
WTForms-Components==0.10.4
zipp==3.4.0
